example/lambda/your-app/requirements.txt
Pillow==7.2.0  
boto3==1.16.52
python-dotenv==0.17.1